Address code review feedback - improve Swiss system
- Fixed Swiss scoring to use 1 point per win (standard Swiss scoring)
- Changed swiss_score from DECIMAL to INT for cleaner scoring
- Improved team filtering performance using Set for O(1) lookups
- Made 144-team requirement more flexible (140-144 range)
- Added clarifying comments for qualification and field assignment
- Updated TODOs with clearer implementation notes

+++ turnier/SQL-Setup-Turnier.sql
@@ -46,7 +46,7 @@ CREATE TABLE IF NOT EXISTS turnier_teams (
     status ENUM('angemeldet', 'bestaetigt', 'abgemeldet', 'disqualifiziert') DEFAULT 'angemeldet',
     teilnehmerzahl INT DEFAULT 2,
     -- Swiss system fields
-    swiss_score DECIMAL(4,1) DEFAULT 0.0,
+    swiss_score INT DEFAULT 0,
     buchholz DECIMAL(8,2) DEFAULT 0.0,
     initial_seed INT DEFAULT 0,
     swiss_qualified BOOLEAN DEFAULT FALSE,
